Create README.md
Python is a very versatile programming language. Python is being used in almost each mainstream technology and one can develop literally any application with it. Let’s see a Python program to convert the currency of one country to that of another country. To use this service, one must need the API key. We will use fixer API to get the live conversion rates and convert the corresponding amount.